fix broken uxplay.cpp logging facility
