+++ lib/raop.h
@@ -32,18 +32,6 @@
 extern "C" {
 #endif
 
-
-/* Define syslog style log levels */
-#define RAOP_LOG_EMERG       0       /* system is unusable */
-#define RAOP_LOG_ALERT       1       /* action must be taken immediately */
-#define RAOP_LOG_CRIT        2       /* critical conditions */
-#define RAOP_LOG_ERR         3       /* error conditions */
-#define RAOP_LOG_WARNING     4       /* warning conditions */
-#define RAOP_LOG_NOTICE      5       /* normal but significant condition */
-#define RAOP_LOG_INFO        6       /* informational */
-#define RAOP_LOG_DEBUG       7       /* debug-level messages */
-
-
 typedef struct raop_s raop_t;
 
 typedef void (*raop_log_callback_t)(void *cls, int level, const char *msg);
